+++ packages/google_adsense/example/lib/main.dart
@@ -10,11 +10,11 @@ import 'package:flutter/material.dart';
 import 'package:google_adsense/experimental/ad_unit_widget.dart';
 import 'package:google_adsense/google_adsense.dart';
 
-void main() {
+void main() async {
 // #docregion init-min
   // Call `initialize` with your Publisher ID (pub-0123456789012345)
   // (See: https://support.google.com/adsense/answer/105516)
-  adSense.initialize('0123456789012345');
+  await adSense.initialize('0123456789012345');
 
   // #enddocregion init-min
   runApp(const MyApp());